Added SV support during simulation
The user can now specify a config file of exactly where SVs should be simulated.
This file will be anyways needed in order to know the exact breakpoints of the SVs.
It supports a few SV types that are documented in the help.
It also supports SVs that don't align with the binning. In case a bin is "split",
the counts in that bin are adapted proportionally.